requirements.txt
Pillow>=10.0
PyPDF2>=3.0
python-docx>=1.1
reportlab>=4.0
